tests: add test for checking guest available MSRs
This test boots a Firecracker uVM and tries to read a set of MSRs from
the guest. The Guest MSR list is compared against a list of MSRs that
are expected when running on a particular host kernel and with a
particular Guest CPU template.
The list is different for each kernel version because Firecracker relies
on MSR emulation provided by KVM. If KVM emulation changes, then the MSR
list available to the guest might change also.
The list is also dependant on CPUID (CPU templates) since some MSRs are
not available if CPUID features are disabled.
Lastly, this tests also checks for MSR values against the baseline. This
helps validate that defaults have not changed due to emulation
implementation changes in the kernel.

Signed-off-by: Alexandru-Cezar Sardan <alsardan@amazon.com>